requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
annotated-types==0.7.0
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.8.0
filelock==3.13.1
fonttools==4.55.8
fsspec==2024.6.1
gdown==5.2.0
gitdb==4.0.12
gitpython==3.1.44
h5py==3.12.1
hdf5storage==0.1.19
idna==3.10
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
lightning-utilities==0.12.0
markupsafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
natsort==8.4.0
networkx==3.3
numpy==2.1.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.29.3
psutil==6.1.1
pydantic==2.10.6
pydantic-core==2.27.2
pyparsing==3.2.1
pysocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyyaml==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
sentry-sdk==2.20.0
setproctitle==1.3.4
six==1.17.0
smmap==5.0.2
soupsieve==2.7
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0+cu124
torch-dct==0.1.6
torchaudio==2.6.0+cu124
torchmetrics==1.6.1
torchvision==0.21.0+cu124
tqdm==4.67.1
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wandb==0.19.5
